A whole bunch of edits that don't really affect DES at all.
Fixed some small memory leak problems discovered with mmgr.
Fixed a bug that caused Seg Faults for some SmallMatrix calculations
  when compiled with ATLAS BLAS.
Updated the documentation to include some of the new features, and the 
  scons installation method.
Added compatibility with Fortran version of BLAS and LAPACK libraries.
Improved the SConstruct files to do a better job of checking for BLAS and
  LAPACK.
Restored LAPACK stegr usage, which required detecting when dstegr fails
  and need to call dstedc instead.
Changed behavior of TMV warning messages.


git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@2828 2422e80f-9519-0410-a819-8e21e7c86b6d