Fixed a bug that was causing icpc -openmp compilations to fail in various non-repeatable ways.
Basically, icpc wasn't handling certain exceptions correctly from within an
opnemp parallel block.  This is a bug in icpc as far as I can tell, since I
am pretty sure my code was compiliant with the openmp standard that just
requires exceptions thrown from within a block be caught in the same block.

Some exceptions seem to be handled correctly, but one of them wasn't.

Anyway, I now deal with the problematic exception in a different way that
seems to be working.  

Also, I added a cool feature where if you are sending debug output to a file,
then it now writes to a different file for each thread, appending _1, _2, etc.
to the file name for all threads other than the master thread.

This was something I wanted to do a while back, but didn't know how.
It turns out that the openmp directive called threadprivate is exactly
what is needed.  


git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@4326 2422e80f-9519-0410-a819-8e21e7c86b6d