I have finally fixed the bug in multishear where it would use much more memory with openmp than it should.
I still don't really understand what was causing the bug, especially why
it only showed up with openmp compilations, but basically what seemed 
to be happening was the memory was getting very fragmented, eventually
90% of the memory allocated by the program would be technically available,
but not large enough for the block sizes I needed.  

The solution was twofold:
1) I now use a shared_ptr<vector<Pixel> > for the underlying data in PixelList
   rather than auto_ptr<vector<Pixel> >.  This let me change the copy 
   constructor's behavior to allow shallow copies, so a second copy of a 
   PixelList shares the same actual vector<Pixel> as the original.
   This is important, since we do push_back(pixlist) on a vector of
   PixelLists.  When vector runs out of space, it needs to copy all its
   elements to new storage.  Now, this operation is both more efficient
   (no actual element copying) and leads to fewer new/delete cycles of the
   PixelList data, so less fragmentation.
2) I now use a pool allocator for the vector<Pixel> stored in PixelList.  
   This forces the program to keep any memory used for these vectors
   earmarked for future PixelList allocations after the memory is nominally
   freed up.  This keeps this memory for getting used by other smaller 
   allocations between the time when I free it (after one section is done)
   and when I need it again (in the next section).

It seems that neither of the above steps works to solve the problem, but both
together do solve the problem.  Go figure.

Anyway, this fix means that the memory usage seems to scale more or less
as const + size^2.  The const is for some memory allocations that are 
needed regargless of what size is used.  But the bulk of the memory usage
is proportionaly to the largest number of galaxies in one of the sections, 
which scales approximately as size^2.



git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@4465 2422e80f-9519-0410-a819-8e21e7c86b6d