It turns out that the boost pool allocator is not that great.
First, while it's faster that what I had for allocating,
it is _way_ slow for deallocating.
So the net result is that it is much slower overall.

And second, because of its strategy of doubling the block size each time,
it ends up allocating up to nearly twice as much as it needs.  So, since
we are often semi-close to the memory limit as it is, this is unacceptable
for our needs.

So I went back to my custom allocator and found out why it was slow,
and fixed it, so it now seems to be working pretty fast.
Not quite as fast as the native new and delete functions probably, 
but close enough that the difference isn't noticeable anymore.


git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@4483 2422e80f-9519-0410-a819-8e21e7c86b6d